add hobby page. each page component now in new file.
