+++ app/src/About/about.sass
@@ -1,10 +1,12 @@
 .page-about
   margin: 0
   border: 0
-  padding: 4rem 0 0
-  img
+  padding: 4rem 0 0 0
+  .img-portrait
     width: 100vw
     max-width: 350px
+  .img-skateboard
+    width: 100vw
   p
     padding: 0
     margin: 1rem
