+++ app/src/About/pages/AboutPage1.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const AboutPage1 = () => {
+  return (
+    <div>
+      <img className='img-portrait' src={require('../images/hire-me.jpg')} />
+      <p>Matt Szabo: Frontend developer looking for new opportunities</p>
+      <p>(Swipe or click the dots below to learn more)</p>
+    </div>
+  );
+}
+
+export default AboutPage1;
