+++ app/src/About/pages/AboutPage2.js
@@ -1,0 +1,24 @@
+import React from 'react';
+
+const AboutPage2 = () => {
+  return (
+    <div className='about-page'>
+        <p>
+          I began learning how to develop for the web at the beginning of May 2016. Since then I have focussed my efforts towards the following skills:
+        </p>
+        <ul>
+          <li>JavaScript with a focus on ES6</li>
+          <li>CSS, Sass, and SVG</li>
+          <li>React</li>
+          <li>Redux</li>
+          <li>Webpack with hot module replacement</li>
+          <li>Node.js</li>
+          <li>Git & GitHub</li>
+          <li>Browser dev tools (Chrome/React/Redux)</li>
+          <li>Responsive & mobile first design</li>
+        </ul>
+    </div>
+  );
+}
+
+export default AboutPage2;
