+++ app/src/About/pages/AboutPage3.js
@@ -1,0 +1,24 @@
+import React from 'react';
+
+const AboutPage3 = () => {
+  return (
+    <div className='about-page'>
+      <h3>My background</h3>
+      <p>
+        I graduated from Computer Science at Monash University and began work in development and operations (DevOps). My most notable experience in this field is my previous position at Shine Technologes where I worked for 5 years. There I supported a large Java application with an SQL backend all hosted on Amazon's AWS architecture. To be effective at the role I utilised and improved on the following skills:
+      </p>
+      <ul>
+        <li>Java</li>
+        <li>SQL</li>
+        <li>UNIX tools</li>
+        <li>AWS cloud management</li>
+        <li>Exceptional problem solving</li>
+      </ul>
+      <p>
+         Another key asset learned during my time in that role was the ability to rapidly learn and understand new technolgies in an ever changing world of programming techniques, frameworks and languages.
+      </p>
+    </div>
+  );
+}
+
+export default AboutPage3
