+++ app/src/About/pages/AboutPage5.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const AboutPage5 = () => {
+  return (
+    <div className='about-page'>
+      <h3>Hobbies</h3>
+      <img className='img-skateboard' src={require('../images/k-grind-nollie-shuvit.jpg')} />
+      <ul>
+        <li>
+          Skateboarding
+        </li>
+        <li>Live music</li>
+        <li>Coding</li>
+        <li>Futsal</li>
+      </ul>
+    </div>
+  );
+}
+
+export default AboutPage5;
